requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
llama-cpp-python>=0.2.0
rank_bm25>=0.2.2
